+++ client/src/game/interfaces/shapes/asset.ts
@@ -1,8 +1,10 @@
+import { AssetId } from "../../../assets/models";
 import type { GlobalPoint } from "../../../core/geometry";
 import type { IShape } from "../shape";
 
 export interface IAsset extends IShape {
     src: string;
+    variants: { name: string | null; assetId: AssetId; width: number; height: number }[];
     svgData?: { svg: Node; rp: GlobalPoint; paths?: [number, number][][][] }[];
 
     get h(): number;