+++ client/src/game/ui/settings/shape/PropertySettings.vue
@@ -308,4 +308,3 @@ button {
     justify-self: flex-end;
 }
 </style>
-../../../systems/properties/helpers