+++ server/src/db/all.py
@@ -8,6 +8,7 @@ from .models.note_tag import NoteTag  # isort: skip
 from .base import BaseDbModel, BaseViewModel
 from .models.asset import Asset
 from .models.asset_rect import AssetRect
+from .models.asset_rect_variant import AssetRectVariant
 from .models.asset_shortcut import AssetShortcut
 from .models.aura import Aura
 from .models.base_rect import BaseRect
@@ -52,6 +53,7 @@ from .signals import *  # noqa: F403
 
 ALL_NORMAL_MODELS: list[type[BaseDbModel]] = [
     AssetRect,
+    AssetRectVariant,
     Asset,
     AssetShare,
     AssetShortcut,