+++ CHANGELOG.md
@@ -10,6 +10,10 @@ tech changes will usually be stripped from release notes for the public
 
 ## Unreleased
 
+### Fixed
+
+-   Asset pick modal butons don't work when clicking on their bottom half
+
 ## [2026.1]
 
 ### Added